backend/analytics-service/requirements.txt
# PayU Analytics Service - Python FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Database (TimescaleDB - PostgreSQL extension)
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.9
timescaledb-psycopg2==2.13.0

# Kafka
aiokafka==0.11.0

# ML / Analytics
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
torch==2.5.0

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0

# Logging & Monitoring
structlog==24.4.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
opentelemetry-exporter-otlp==1.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
